Contracts refactor phase 3: extract ContractsDataService
Moves the dashboard normalization/merge pipeline out of
ContractsPageComponent into a new ContractsDataService: turning registry
entries and indexer responses into ContractDashboardEntry rows
(localEntryToDashboard, indexerSummaryToDashboard), merging local +
indexer sources, participant resolution (local/template/self-custody/
indexer), next-action-label logic, alias resolution, and the indexer
covenant-resolution cluster (fetchIndexerCovenant and friends).

Also extends CovenantTemplateService with pubkeyToAddress and the
extractTemplate*/templateForIndexerName helpers, and ContractDisplayService
with getTemplateDisplayName/getContractTypeLabel — pure template/crypto
and label helpers the dashboard pipeline needs that are also used by
not-yet-extracted lookup/detail/action code, so they belong alongside
the other shared logic from phase 2 rather than duplicated.

refreshContractStatuses stays in the shell rather than moving: it writes
directly through shell-owned signals (allRegistryContracts,
registryContracts) via updateRegistryContract, so it's orchestration,
not stateless normalization.

Every call site outside the dashboard load path (lookup/import, detail,
action execution — none extracted yet) keeps working unchanged through
thin same-signature wrapper methods that delegate to the new services.

No template changes; loadContracts()'s body now calls into
ContractsDataService instead of shell-private methods.

Shell: 6,938 -> 6,208 lines (.ts).
